Move getSsmParameter to tests/common
